Initial commit
I was going to make a commit earlier than this, but I had forgotten to
do it...

I started this project a couple days ago and was spending 2hrs after
work, but today I have been working on it for over 12hrs.

I think it works now. The 3 examples I have of the renderer work
flawlessly (at least now).